PULSEDEV-32102 Add support for large LightGBM datasets (#37)
* PULSEDEV-32102 Add support for large LightGBM datasets

- Add support to large datasets in LightGBM
  - Use the new LightGBM API: allocate with long parameter (not int)
  - Change the product of integer dimensions of the buffer to a long
- Replace microsoft's lightgbmlib build by our own
- Update LightGBM to a very recent commit in the master with fast
  prediction support - fc79b36 (LightGBM's #2992)
- Tag this as lightgbmlib=2.3.160 for a lack of better choice
    This is > microsoft's 2.3.150, although unfortunately it diverges from
    microsoft's lightgbmlib >=2.3.150 releases

* Fix build without microsoft lightgbmlib

- Split openml-lightgbm module into 2:
  - lightgbm-builder: builds lightgbm + lightgbmlib + installs it locally
  - openml-lightgbm-provider: The provider which requires lightgbm-builder's artifacts
- Parent openml-lightgbm module selects the LightGBM build versions

* Move artifacts into lightgbm-builder (will break)

* Streamlined build. Perform modifications in moved files.

* Build working locally with dependency across the two LGBM submodules.

- Update pom's
- Update readme to explain how to compile

* Address Sheng comments

* Move files (will break)

* Change moved files

* Support new custom LIGHTGBM_REPO_URL parameter in make-lightgbm.

* Fix large dataset copy index type for SWIG copy

* Move modules

* Modify moved modules

* Move module

* Changed moved modules

* Working build with reorganized groupIds

* Working build with more organized groupIds

* Bump lightgbmlib version to 2.3.190

* [fix] Change setitem rowOffset product computation to long.

* Rename VERSION to COMMIT_REF

+++ openml-lightgbm-meta-module/openml-lightgbm/src/main/java/com/feedzai/openml/provider/lightgbm/SWIGTrainResources.java
@@ -93,7 +93,7 @@ public class SWIGTrainResources implements AutoCloseable {
 
         logger.debug("Allocating SWIG train data array.");
         // 1-D Array in row-major-order that stores only the features (excludes label) in double format:
-        this.swigTrainFeaturesDataArray = lightgbmlib.new_doubleArray(numInstances * numFeatures);
+        this.swigTrainFeaturesDataArray = lightgbmlib.new_doubleArray(((long)numInstances) * ((long)numFeatures));
         // 1-D Array with the labels (float32):
         this.swigTrainLabelDataArray = lightgbmlib.new_floatArray(numInstances);
